Done with renaming movieUtils and correcting hook implementation of playlists
